Add files via upload
In this paper, we compare growth functions of different bases of commutator subgroups of free groups. Of the bases that we consider, the geodesic Schreier basis appears to be the fastest. In connection with groups, we consider a basis for the free Lie commutator subalgebra. In contrast with groups, the growth of a basis for a free Lie algebra does not depend on the choice of basis.