Add files via upload
In this paper, we give a probabilistic existence proof of the countable
random graph R and discuss some of its graph theoretic properties. A
more constructive existence proof will also be given. We will see that the
graph is indestructible, in that a finite number of changes to its vertices or
edge set result in a graph isomorphic to R. Furthermore, we will explore
the automorphism group of R, and will show that this group is simple and א0
has cardinality 2 .